requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0
tqdm>=4.64.0
numpy>=1.22.0
matplotlib>=3.5.0
peft>=0.4.0 